+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.2.0 -- YYYY-mm-dd
 
+* Add generation of pure variants of translation facilities to `defpass`.
+* Change `{Xlate,descend*}A` names to drop the `A`; applicative is probably the more common case.
 * Fix bug in testing for `Traversable` instance.
 * Generate documentation for generated types/functions.
 * Requires template-haskell >=2.18, and therefore GHC 9.2.1